feat(mariadb): disable ssl client requirement
this disables the required ssl connection for a mariadb client to the server.
most of our mariadb servers are on the same host and in the same docker network as
the application, so we don't force the client to connect to a ssl enabled mariadb server.

this default config applies to every php inherited image, as they mostly adds the
mariadb client binaries in the project.

we'll remove this after every mariadb server offers ssl in our network and on
the machine of the developers in the future.